% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{calculate}
\alias{calculate}
\title{Calculate overlap statistics}
\usage{
calculate(input, universe = NULL)
}
\arguments{
\item{input}{output of \code{\link{process_input_text}} or
\code{\link{process_input_seurat}}}

\item{universe}{optional: number of genes in universe; default calculate from
\code{gs_genes}}
}
\value{
\code{stats} tibble: overlap statistics

\code{matches} list: names: annotations vector: matched genes
}
\description{
Calculate overlap statistics
}
\examples{
\dontrun{
input <- process_input_text("FCN1 0.1 FTL 0.8 CLU 0.05")
calc <- calculate(input)
}
}
\keyword{internal}
