% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{calculate_post}
\alias{calculate_post}
\title{Finish calculating overlap statistics}
\usage{
calculate_post(stats_pre, input_size, universe)
}
\arguments{
\item{stats_pre}{value from \code{\link{calculate_pre}}}

\item{input_size}{number of genes in input}

\item{universe}{number of genes in universe}
}
\value{
tibble: overlap statistics
}
\description{
Finish calculating overlap statistics
}
\examples{
\dontrun{
input <- process_input_text("FCN1 0.1 FTL 0.8 CLU 0.05")
calc_pre <- calculate_pre(input)
calc <- calculate_post(calc_pre$stats_pre, nrow(input), 10000)
}
}
\keyword{internal}
