% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{calculate_pre}
\alias{calculate_pre}
\title{Begin calculating overlap statistics}
\usage{
calculate_pre(input)
}
\arguments{
\item{input}{output of \code{\link{process_input_text}} or
\code{\link{process_input_seurat}}}
}
\value{
\code{stats_pre} tibble: overlap statistics (incomplete)

\code{matches} list: names: annotations vector: matched genes
}
\description{
Begin calculating overlap statistics
}
\examples{
\dontrun{
input <- process_input_text("FCN1 0.1 FTL 0.8 CLU 0.05")
calc_pre <- calculate_pre(input)
}
}
\keyword{internal}
