% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{compute}
\alias{compute}
\title{Perform computations on imported and processed data}
\usage{
compute(input, universe = NULL, save = NULL)
}
\arguments{
\item{input}{output of \code{\link{process_input_text}} or
\code{\link{process_input_seurat}}}

\item{universe}{number of genes in universe}

\item{save}{optional: path to save overlap statistics as \code{.csv}}
}
\value{
\code{stats} tibble: overlap statistics

\code{matches} list: names: annotations vector: matched genes
}
\description{
Perform computations on imported and processed data
}
\examples{
\dontrun{
input <- process_input_text("FCN1 0.1 FTL 0.8 CLU 0.05")
results <- compute(input)
}
}
