% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{explore_annotation}
\alias{explore_annotation}
\title{Find an annotation's associated gene sets and genes}
\usage{
explore_annotation(annotation, genes = NULL)
}
\arguments{
\item{annotation}{annotation to explore}

\item{genes}{optional: genes to match, or (default) all}
}
\value{
\code{"names"} vector: gene set names

\code{"genes"} vector: genes
}
\description{
Find an annotation's associated gene sets and genes
}
\examples{
\dontrun{
anno_assoc <- explore_annotation("Carcinogen")
}
}
\keyword{internal}
