% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{get_data}
\alias{get_data}
\title{Get processed database}
\usage{
get_data()
}
\value{
\code{gs_genes} list: names: gene set names vector: genes

\code{gs_info} tibble: gene set information

\code{genes} vector: list of genes
}
\description{
Get processed database
}
\examples{
\dontrun{
get_data()
}
}
\keyword{internal}
