% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_annotations}
\alias{process_annotations}
\title{Process annotations}
\usage{
process_annotations()
}
\value{
\code{gs_annos} tibble: gene sets and annotations

\code{annos} vector: annotations
}
\description{
Process annotations
}
\examples{
\dontrun{
anno_proc <- process_annotations()
}
}
\keyword{internal}
