% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_database}
\alias{process_database}
\title{Process database}
\usage{
process_database(data)
}
\arguments{
\item{data}{database}
}
\value{
\code{gs_genes} list: names: gene set names vector: genes

\code{gs_info} tibble: gene set information

\code{genes} vector: list of genes
}
\description{
Process database
}
\examples{
\dontrun{
data <- import_msigdb(msig_path)
data_proc <- process_database(data)
}
}
\keyword{internal}
